Use Modalize HOC for AddFrameworkButton and CloneFrameworkButton
