--- src/views/Project/Details/Frameworks/AddFrameworkForm/styles.scss
@@ -1,29 +1,0 @@
-@import '~base-scss/base';
-
-.add-analysis-framework-form {
-    display: flex;
-    position: relative;
-    flex-direction: column;
-    height: 100%;
-
-    .title {
-        margin-bottom: $spacing-medium;
-    }
-
-    .description {
-        :global .input {
-            resize: none;
-        }
-    }
-
-    .action-buttons {
-        display: flex;
-        justify-content: flex-end;
-        margin-top: $spacing-medium;
-
-        button {
-            margin-left: $spacing-medium;
-        }
-    }
-}
-
