+++ src/views/Project/Details/Frameworks/FrameworkDetail/CloneFrameworkModal/styles.scss
