--- src/views/Project/Details/Frameworks/FrameworkList/AddFrameworkButton/styles.scss
@@ -1,11 +1,0 @@
-@import '~base-scss/base';
-
-$width-add-framework-modal: 480px;
-
-.add-framework-modal {
-    width: $width-add-framework-modal;
-
-    .modal-body {
-        padding: $spacing-large;
-    }
-}
